refactor: centralize API error handling with `genericApiErrorHandler` across modules
