+++ apps/admin-client/src/features/admins/api/index.ts
@@ -4,10 +4,9 @@ import {
   adminsListSchema,
 } from '#/features/admins/schemas';
 import type { AdminListInputSchema } from '#/features/admins/schemas';
-import { authFetch } from '#/lib/auth-fetch.ts';
+import { authFetch, genericApiErrorHandler } from '#/lib/auth-fetch.ts';
 import { env } from '#/env.ts';
 import { createSearchParams } from '#/lib/utils.ts';
-import { apiErrorSchema } from '#/lib/schemas.ts';
 
 /**
  * Retrieves a list of admin users from the server.
@@ -43,14 +42,7 @@ export const getAdminsList = createServerFn()
       },
     );
 
-    if (!response.ok) {
-      const error = await response.json();
-      const validationSchema = apiErrorSchema.safeParse(error);
-      if (!validationSchema.success) {
-        throw new Error('Failed to fetch admins');
-      }
-      throw new Error(validationSchema.data.detail);
-    }
+    await genericApiErrorHandler(response);
 
     const responseData = await response.json();
