+++ apps/admin-client/src/features/parkings/api/place-types.ts
@@ -8,9 +8,8 @@ import {
   updatePlaceTypeInputSchema,
 } from '#/features/parkings/schemas';
 import { createSearchParams } from '#/lib/utils.ts';
-import { authFetch } from '#/lib/auth-fetch.ts';
+import { authFetch, genericApiErrorHandler } from '#/lib/auth-fetch.ts';
 import { env } from '#/env.ts';
-import { apiErrorSchema } from '#/lib/schemas.ts';
 
 /**
  * Fetches a list of place types from the server based on the provided input data.
@@ -39,9 +38,7 @@ export const getPlaceTypes = createServerFn()
       },
     );
 
-    if (!response.ok) {
-      throw new Error('Failed to fetch place types');
-    }
+    await genericApiErrorHandler(response);
 
     const responseData = await response.json();
 
@@ -92,14 +89,7 @@ export const createPlaceType = createServerFn()
       }),
     });
 
-    if (!response.ok) {
-      const error = await response.json();
-      const validationSchema = apiErrorSchema.safeParse(error);
-      if (!validationSchema.success) {
-        throw new Error('Failed to create place type');
-      }
-      throw new Error(validationSchema.data.detail);
-    }
+    await genericApiErrorHandler(response);
 
     const responseData = await response.json();
 
@@ -139,14 +129,7 @@ export const updatePlaceType = createServerFn()
       },
     );
 
-    if (!response.ok) {
-      const error = await response.json();
-      const validationSchema = apiErrorSchema.safeParse(error);
-      if (!validationSchema.success) {
-        throw new Error('Failed to update place type');
-      }
-      throw new Error(validationSchema.data.detail);
-    }
+    await genericApiErrorHandler(response);
 
     const responseData = await response.json();
 
@@ -196,14 +179,7 @@ export const deletePlaceType = createServerFn()
       },
     );
 
-    if (!response.ok) {
-      const error = await response.json();
-      const validationSchema = apiErrorSchema.safeParse(error);
-      if (!validationSchema.success) {
-        throw new Error('Failed to delete place type');
-      }
-      throw new Error(validationSchema.data.detail);
-    }
+    await genericApiErrorHandler(response);
 
     const responseData = await response.json();
